Define Bundler::Override in bundler/override.rb
Keeps "load all lib" more simple.

Note: Actually, the two workhorses of the plugin are `DslPatch` and
`DependencyPatch`. `Bundler::Override` itself is only a registry shared
by both.

To `require_relative "../override"` from each `bundler/override/*.rb`
would express the actual dependency more precisely. But to require the
shared code early and once is also good.
